Create is it a leap year
This code was for a question I completed on HackerRank. The question was Given a year, determine whether it is a leap year. If it is a leap year, return the Boolean True, otherwise return False.

Note: I was given that the year can be evenly divided by 4, is a leap year, unless: The year can be evenly divided by 100, it is NOT a leap year, unless: The year is also evenly divisible by 400. Then it is a leap year.