Create Find the runner up score
this was a HackerRank question. The question was asking, Given the participants' score sheet for your University Sports Day, you are required to find the runner-up score. You are given  N scores. Store them in a list and find the score of the runner-up.

basically find the second best score