Create merge tool
this was a Hacker Rank question. in simple terms the question was- given a single string of letters, Split the string into len(string)/k equal parts of length k. make sure that each of your end strings don't have any repeating letters. 

note: k is given in the input as k