Create forcast of rain and temp
The API Key should be replaced with your API key that you get from Weather API. 

this code is part of a passion project I have, using an API given by Weather API, I am able to get data about the percent of rain, the average temperature, and the total precipitation or rain in mm that day, for the next 7 day. This is the part of the code I am going to use to make a drip irrigation system.